a2a-python-sdk/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "fox-a2a-sdk"
version = "1.0.0"
description = "A2A Framework Python SDK - Agent-to-Agent communication over gRPC"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.62.0",
    "grpcio-tools>=1.62.0",
    "protobuf>=4.25.0",
    "PyJWT>=2.8.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "black", "mypy"]

[tool.setuptools.packages.find]
where = ["."]
include = ["a2a*"]
